feature 10: show hourly request histogram
Drawing *** or similar on console seemed not feasible, learned that one
can use javafx to draw such charts, so went with this solution.
Also tried JfreeChart library but it seems to be unmaintained and
integration wasn't as straight forward as I had hoped.
Sadly it seems javaFX wants to be the starting point of an application,
could not simply call it from my other main method. (it didn't find the
file although it was provided in args). Since showing this/any UI will
keep the app running, it would also disturb the execution time info
(feature 8), therefore I made it to a separate task which can be launched
via 'gradle histogram', which will use timing.log as argument. If the
file name&path are different, one has to modify the gradle file.